refactor: rename fitSetting.ts to fitSettingTab.ts
